[refactor] #4 - ReLU, Sigmoid 클래스 trainer/activation.py에서 model/layers.py로 이동
