+++ model/layers.py
@@ -1,0 +1,32 @@
+import numpy as np
+
+class Relu:
+    def __init__(self):
+        self.mask = None
+    
+    def forward(self, x):
+        self.mask = (x <= 0)
+        out = x.copy()
+        out[self.mask] = 0
+        
+        return out
+    
+    def backward(self, dout):
+        dout[self.mask] = 0
+        dx = dout
+        
+        return dx
+    
+class Sigmoid:
+    def __init__(self):
+        self.out = None
+        
+    def forward(self, x):
+        out = 1 / (1 + np.exp(-x))
+        self.out = out
+        
+        return out
+    
+    def backward(self, dout):
+        dx = dout * (1 - self.out) * self.out
+        return dx
